+++ packages/insomnia/src/main/ipc/electron.ts
@@ -37,6 +37,7 @@ export type HandleChannels =
   | 'generateCommitsFromDiff'
   | 'generateMockRouteDataFromSpec'
   | 'getAuthHeader'
+  | 'getOAuth2Token'
   | 'getExecution'
   | 'getLocalStorageDataFromFileOrigin'
   | 'git.abortMerge'
@@ -80,6 +81,7 @@ export type HandleChannels =
   | 'grpc.loadMethods'
   | 'grpc.loadMethodsFromReflection'
   | 'grpc.writeProtoFile'
+  | 'initNewOAuthSession'
   | 'insecureReadFile'
   | 'insecureReadFileWithEncoding'
   | 'installPlugin'
