+++ packages/insomnia/src/ui/components/modals/oauth-authorization-status-modal.tsx
@@ -2,7 +2,7 @@ import React, { type FC, useEffect, useRef, useState } from 'react';
 
 import { useDefaultBrowserRedirectActionFetcher } from '~/routes/auth.default-browser-redirect';
 
-import type { OAuth2AuthorizationStatusType } from '../../../network/o-auth-2/constants';
+import type { OAuth2AuthorizationStatusType } from '../../../main/network/o-auth-2/get-token';
 import { invariant } from '../../../utils/invariant';
 import uiEventBus, { OAUTH2_AUTHORIZATION_STATUS_CHANGE } from '../../event-bus';
 import { Modal, type ModalHandle } from '../base/modal';
