+++ packages/insomnia/src/ui/components/settings/general.tsx
@@ -1,8 +1,6 @@
 import React, { type FC, Fragment } from 'react';
 
 import { useRootLoaderData } from '~/root';
-import { clearOAuthWindowSessionId } from '~/ui/spawn-oauth-window';
-
 import {
   EditorKeyMap,
   MAX_EDITOR_FONT_SIZE,
@@ -226,7 +224,7 @@ export const General: FC = () => {
         />
         <button
           className="pointer h-(--line-height-xs) rounded-md border border-solid border-(--hl-lg) px-(--padding-sm) hover:bg-(--hl-xs)"
-          onClick={clearOAuthWindowSessionId}
+          onClick={window.main.initNewOAuthSession}
         >
           Clear OAuth 2 session
         </button>
